CHEM-010: rename files
