+++ src/redux/action-types/applicationsActions.ts
@@ -1,0 +1,3 @@
+const enum ApplicationsActions {
+    ADD_APPLICATION = 'ADD_APPLICATION',
+}
