+++ src/redux/reducers/applicationsReducer.ts
@@ -1,0 +1,5 @@
+const applicationsReducer = (state, action) => {
+
+}
+
+export default applicationsReducer;
