+++ src/redux/reducers/index.ts
@@ -1,8 +1,10 @@
 import { combineReducers } from 'redux';
+import applicationsReducer from './applicationsReducer';
 import chemicalApplicationReducer from './chemicalApplicationReducer';
 
 const reducers = combineReducers({
-    chemicalApplication: chemicalApplicationReducer
+    chemicalApplication: chemicalApplicationReducer,
+    applications: applicationsReducer
 });
 
 export default reducers;
